Unified upload endpoints (#511)
This PR introduces a number of improvements to upload logic:
- removes artifact-specific upload endpoints and instead introduce two
centralized upload endpoints. This is better because identical endpoints
would have to be replicated for future pickle uploads, and other uploads
(e.g. logs, see below)
- streamlines the storage plug-in interfaces so that the same behavior
happens regardless of the the plugin (client asks for upload location,
puts to upload location, gets from download location which always
redirects to the actual location)
- The two upload endpoints are 1/ give me a location where I can PUT my
payload, 2/ give me the stored bytes via a redirect
- The `LocalStorage` plug-in adds two endpoints that essentially
directly replace the S3 PUT and GET endpoints so that the interface and
behavior is identical with `S3Storage`

### Additional reason for centralized upload endpoints

The original impetus for this whole upload work was to fix matplotlib
support for deployed API servers. Unlike Plotly, matplotlib does not
have a JSON/Typescript API, which means that figures need to be
persisted to image (e.g. png) and served to the UI. With local servers,
we do this by writing the image to local disk and letting the server
read it from there and serving it to the UI. For deployed servers this
does not work as the local filesystem is not persistent.
Instead we have to store the image on S3 and let the server serve them
from there.

Essentially what needs to happen is that the `MatplotlibFigureValueView`
React component asks the server for the full artifact bytes (not just
the JSON summary) and displays it. That can work well with the new
upload system.

Now the issue is, often times, these Matplotlib figures are not
standalone artifacts. They are fields in a dataclass or other container
types. How does this get serialized and served to the UI? Of course we
could store the entire image binary in the JSON summary of the
dataclass, but that seems unreasonable and will break the capacity
limits of the DB.

A solution to that is to let the image be stored as a binary blob on S3
(not an artifact, just a single blob) and let the JSON summary of the
`matplotlib.figure.Figure` type contain a pointer to that blob (e.g. a
blob id). When the UI encounters an artifact or a field in a container
type of type `matplotlib.figure.Figure`, it will recognize the blob
reference and query the API for a download location.

This solution can also be used for `PIL.Image.Image` types (type seen in
the YOLO pipeline), or other image types.

This means that in addition to the two existing upload namespaces
(`artifacts` and `futures`), we now have a third (e.g. `blobs`) that can
be used for arbitrary binary blobs that cannot fit in a JSON summary in
the DB.

Therefore, instead of replicating three times the same endpoints it is
better to have centralized upload endpoints under the `/api/v1/uploads`
namespace. The endpoints are then parametrized by a `namespace` (one of
`artifacts`, `futures`, `blobs`) and a `key` (`artifact_id`, `run_id`,
and some UUID to identify blobs).

---------

Co-authored-by: augray <augray@users.noreply.github.com>
Co-authored-by: tscurtu <tudor@sematic.dev>