Warn on Python migration error (#549)
Currently when a Python migration fails, the exception is swallowed and
a revert is immediately executed, creating a bad experience.

This updates the migration logic to also inform the user what the
original exception was.

Also, an existing migration script is edited to permit accepting the new
version of the settings file and completing successfully, instead of
only accepting the previous version that had to be migrated away from.