Use custom SA for Ray (#570)
Addresses #568 

In the same way we want Ray workers to have access to the same code
dependencies as the associated Sematic func, we also want them to have
access to the same permissions. This means that if a custom SA is being
used for Sematic workers, it should be used for Ray workers as well.

Testing
-------
In an environment using a custom SA for Sematic workers, launched a Ray
cluster using `with RayCluster`. Looked at the pod SA for the workers
and the head to confirm that they had the custom SA set.

---------

Co-authored-by: Josh Bauer <josh@sematic.dev>