Add Re-run and Slack documentation (#563)
Adds user-facing documentation for the Slack integration and for the
Re-run feature.

Removes the "beta" label from the Dashboard buttons for the re-run
feature.

Also opportunistically enables a button in the Dashboard to copy the
current Run URL, as all prerequisite deep linking work has been
completed, and because this code was adjacent to other touched code.