Fix storage/logging regression (#580)
Worker/driver jobs are not getting settings around storage transferred
to them, but try to use S3 storage to write logs. The proper fix is to
integrate logging into the full storage plugin system so the
workers/drivers can be agnostic of where they are writing the logging
file bytes to. In the meantime, we need to keep the `main` branch in a
releasable state. See #579 .

Note: I'm not sure at exactly what point this regression was introduced,
but it will be masked if your `settings.yaml` configures an
`AWS_S3_BUCKET`. Why? Because when your settings file configures that,
your user settings (including that configuration) get uploaded as env
vars for the worker/driver.

Testing
--------
Submitted a job from a machine where `AWS_S3_BUCKET` was not configured
in `settings.yaml`, but to a server where it *was* configured. Confirmed
that logs showed up for the runs.