Refactor in-app links to use react router (#585)
Instead of using the default `<Link>` component from MUI for in-app
links, change them to use Links from `react-router`, so when these links
are clicked, there will be no full page reload.

The component is implemented as a composition of the `<Link>` component
from `MUI` and the one from `react-router`. So the MUI styles will still
apply.


If interested, here is how React-router's `<Link>` changes the browser
URL without reloading the page


https://github.com/remix-run/react-router/blob/2b843a6739c6807e683de9097e20909e78f0a1bd/packages/react-router-dom/index.tsx#L446-L473