Logs use new storage code (#608)
This PR makes sure the storage code that supports log ingestion and
streaming is implemented on all storage plugins, and that log files are
ingested using pre-signed URLs as opposed to directly to S3.

After this PR merges, workers will no longer need the access to the S3
bucket.

Test run with logs
https://emmanuel.dev-usw2-sematic0.sematic.cloud/pipelines/sematic.examples.testing_pipeline.pipeline.testing_pipeline/93e043e7137c4ffa9036e3a977775c48#tab=logs&run=01153cf6de0c4cc19785f7b35dae9cc8