[3/3] Migrate notes to point to user ID instead of email (#602)
This PR migrates the `notes` table and `Note` model to deprecate
`author_id`, which points to `users.email`, in favor of a new `user_id`
field pointing to `users.id`.

The SQL migration includes the creation of a temporary table to
circumvent SQLite limitations (namely the impossibility to use `JOIN` in
an `UPDATE` query).

The PR also introduces an `HasUserMixin` mixin and centralizes storage
of mixins in `sematic/db/models/mixins`.

---------

Co-authored-by: augray <augray@users.noreply.github.com>