Log message to be used for pipeline stats (#620)
Some users would like to be able to use their monitoring infra to create
dashboards around various pipelines. A simple way for us to enable this
for them is to add a log message that contains the relevant info about a
pipeline's terminal state. Any users who rely on this should do so at
their own risk: this is not a public API and is subject to change at any
time. However, it can be useful if people are willing to accept that it
may change/disappear.

Testing
--------
Deployed and ran a couple pipelines; one of which was known to fail:

Sample messages:
```
2023-03-02 18:53:53,050 [INFO] sematic.api.endpoints.resolutions: Remote resolution df55c4cccf5c4573b8812ca882deef00 for pipeline sematic.examples.testing_pipeline.pipeline.testing_pipeline terminated with root run in state NESTED_FAILED. The root run had tags: ["example", "testing"]
2023-03-02 18:54:17,548 [INFO] sematic.api.endpoints.resolutions: Remote resolution 34df7c05dedd40cd99d4d3ff50c4b953 for pipeline sematic.examples.testing_pipeline.pipeline.testing_pipeline terminated with root run in state RESOLVED. The root run had tags: ["example", "testing"]
```