`Image` type (#623)
This PR adds support for returning arbitrary images in Sematic
functions, and having them show up in the UI.

This requires a change to the JSON summary generation API, so that
arbitrary binary blobs can be stored.

Why not have the UI simply get the artifact's binary payload and display
that?
Because often times, the image will be part of a container (e.g.
dataclass, list, tuple, dict), which will constitute the entire
artifact. So instead, we store just the image part and the UI knows to
fetch it.

This mechanism can be used in the future to display lists and dataframes
of arbitrary sizes, if we can figure out a way to stream the binary blob
instead of fetching it all at once.
It can also be used to replace our current broken matplotlib support.

Current type API:
```python
@sematic.func
def image() -> Image:
    image = Image(bytes=image_bytes)
    # or
    image = Image.from_file("/path/to/file")
    return image
```

Returning a `List[Image]`:

![Screenshot 2023-03-02 at 4 10 23
PM](https://user-images.githubusercontent.com/429433/222596693-d2c95828-8c26-4f4f-8685-4da0349ab56a.png)