Separating common package and use absoulte import paths (#651)
1. Separate a "common" package to hold common structs referenced by
`main` and `test` packages.

This package is a typescript reference project, which means the
depending package will not rebuild it if there is nothing changed after
the last build. Potentially accelerate compile time.

2. Moved commonly used typings for Types visualization to `common`
package, updated references across
3. When updating the above references, also updated involved files to
use the absolute package import path (eliminate `../` style relative
import paths ).
4. In `tsconfig` of the test package, add base paths, so it can resolve
the absolute import path used by the `main` package.
5. Manually create the DevServer used by Cypress, so we have a chance to
change its webpack config to have the test package also recognize the
absolute paths in the `main` package.
6. Have a `eslint-config` package to store commonly used eslint config.
(It is easier to extend if it is a package)


Tips for reviewing:
This PR is large, but the files under `packages/main/src` are only
import path changes.