Fix S3 signatures by specifying region and signature version (#652)
The `Image` type PR #623 added caching to the storage `/data` endpoint
and the S3 pre-signed URL.
The caching was always a little risky and made the assumption that both
URLs (`/data` and S3) would be cached and expire at the same time, which
is fickle. This PR removes S3 caching at this time.

Additionally, not specifying the bucket region and signature version to
the boto3 client seems to lead to unstable URL signatures. This PR sets
both region and signature version. The region needs to be set in the
Helm value file.