Allow caching dev requirements for server build (#659)
Since the `pip install` of the Sematic wheel is one step in the image
build, Docker is not able to cache the install of Sematic's
requirements. Requirements installs go in the same layer as whatever
local changes have been made. This makes the build & push slower than
they need to be, and also makes it so k8s nodes aren't able to
effectively cache a layer for the image with deps installed. During
iterative development and deployment, the wasted time can add up. This
makes a couple changes to the Dockerfile that will make it possible
during dev to cache the requirements install in a different image layer.
This is meant to be used in conjunction with a change in the "serve-dev"
development deployment script, which will be able to leverage the
caching.