Decrease CloudResolver max poll interval to 1 min (#655)
Decreases the `CloudResolver` max polling interval from 10 min to 1 min.
This will mean that the expected reaction time for progressing on DAG
execution decreases from 5 min to 30 s.

The advantage is that after having waited for long-running runs to
complete, the Resolver will continue scheduling new, unblocked runs,
after waiting a less amount of extra time.

The disadvantage is more logs being outputted by the Resolver, but these
are not viewable yet by the users anyway in the Cloud execution workflow
that is improved by this PR.