Make notifying about updates not fatal on failure (#676)
Whenever a resolver makes updates to graph/pipeline state, it tries to
send out a notification that it has done so. This helps the UI refresh
promptly on changes. However, sometimes the socketio server is unable to
fulfill the request to broadcast, in which case the resolver would
crash. Since this notification is for UI purposes only, it should not be
fatal to a resolution. This PR makes it so that is the case.