Fixes bug when displaying duration for cloned run (#682)
For cloned runs, we fill neither the `failed_at` or `resolved_at`
fields. In order to display duration correctly, when we see the run has
"original_run_id" field, we hardcode the running duration to "<1s"

This will not fix the issue if `failed_at` or `resolved_at` fields are
null for failed or canceled runs. That will be investigated separately.