Use URL parameter to construct redirects (#688)
Depending on where the client is making a request (in-cluster, browser,
local dev box), the host and scheme of a redirect URL are different.

If we controlled a single deployment we could make sure the correct
headers are forwarded so as to reconstruct the origin host, but because
we don't control every deployment environments, we must use some other
mechanism.
In this PR, an `origin` URL parameter is added to queries to store and
retrieve stored data.