Make local Server run with Python 3.10 (#694)
The local Server does not work when installed and started from the wheel
for Python 3.10 only. Running from Bazel is unaffected.

The webserver was not accepting connections at all. The fix was to apply
the `gevent` monkey patch as the first thing on importing the `sematic`
module.

Tested locally in a virtualenv by starting the Server and submitting a
pipeline.