fix CRA cross-package compile issue (#693)
There was an issue that if the `main` package imports a ts file from the
`common` package, and the ts file in the `common` package has an import
with an absolute path ("src/file.ts" instead of "../file.ts"), the
compilation of the `main` package (AKA **C**reate **R**eact **A**pp
application) will fail.

This is because the default webpack configuration of CRA in the `main`
package, when resolving modules, would not consider paths other than the
`main` package.

The solution is:
1. Define in the `tsconfig` of the `main` package that it references the
`common` package. (So IDE intellisense will work)
2. Use
[react-app-rewired](https://www.npmjs.com/package/react-app-rewired) to
extend the base webpack config of CRA. And use
[tsconfig-paths-webpack-plugin](https://www.npmjs.com/package/tsconfig-paths-webpack-plugin)
to ensure that the CRA webpack will understand module resolution the
same way as `tsconfig` does.


This PR also configs the `test` package to understand the module
resolution of files under the `common` package. There is a new unit test
to show that the Cypress component tests can also handle a test in the
`common` package.