Ensure DB schema dump (#704)
Recently some migration files were omitted from the `sqlite` schema file
dump. The effect is that fresh local installations can be created in
inconsistent states, that might result in un-applicable subsequent
migrations.

To avoid this situation, this PR makes it so the schema is dumped in the
pre-commit Make target, which we realistically always run.

Also, this PR includes a dump from a fresh and empty copy of the schema.
During development, we might apply migrations which do not end up in
main (most often due to changes in their timestamp name component). The
result is that developers' local `sqlite` DBs are polluted with dev
migrations, which do not actually exist in the codebase.