Use react hooks in story components (#712)
Back in [PR 702](https://github.com/sematic-ai/sematic/pull/702) I
introduced stories for the RunTree component. There I used a class
component for the story because I had trouble using functional
components and hooks.

Later investigation indicates that we saw the issue with hooks because,
in Storybook's setup, we had two copies of React (despite the same
version). The official React manual
[here](https://legacy.reactjs.org/warnings/invalid-hook-call-warning.html#duplicate-react)
explains that.

My solution is to reimport the same react hooks used in the `common`
package and consume them in the storybook package. This approach is
easier to manipulate than symbolic linking the react dependency from the
storybook, which will also have the bundler (webpack) treat the two
references to the react npm module from `common` and `storybook`
packages the same.

This PR also converted the class component into the functional component
in the storybook for the sake of being streamlined.