Ensure different users can rerun Resolutions (#714)
This fixes a regression where pipeline Resolutions could not be rerun by
different users compared to the original submitters:
```
2023-03-30T23:44:29.471676245Z stdout F 2023-03-30 23:44:28,929 - ERROR - sematic.api_client: Server returned 400 for PUT http://sematic-server/api/v1/resolutions/46ae1ecf9a0243c580483140c68ba362: {"error": "Cannot update user_id of resolution 46ae1ecf9a0243c580483140c68ba362 after it has been created. Original value: '146161c5594c49c286dd6fb505761d3c', new value: '650ac18ea09f4c36ad69f752e2ff4620' (will not be used)"}
```

This regression actually disabled a previously-existing escalation of
privileges vulnerability, where a reran Resolution would execute using
the original submitter's API key, instead of the API key of the user who
triggered the rerun.

When a Resolution was cloned, it had its user id, and the user id of the
root Run updated to the new user's id, but the pod that was created
still had the original user's API key in its env variables. The result
was that this API key was used to source the identity of the user, which
was added to all calls from the Resolver to the Server. On these calls,
the Server re-overwrote the user id on the Resolution and on all Runs.

This PR makes it so that the user id and API key are always set on the
new Resolution, which will end up in the Resolver pod, and correctly use
the identity of the rerun submitter.

Also, several other minor bugs are fixed and improvements are made to
the workflows exercised in debugging this issue.

## Testing

Added a unit test case to cover the originally-reported issue, and
several unit tests to test authentication and user identity override on
the touched endpoints.

Tested re-running the pipeline (the original run was triggered by
another user):
- in the cloud using the Dashboard as trigger -
[result](https://tudor.dev-usw2-sematic0.sematic.cloud/runs/f4ae60797aed4024be0150d97188f7ff)
- in the cloud using the CLI as trigger -
[result](https://tudor.dev-usw2-sematic0.sematic.cloud/runs/0c351fe6739147c48efbef7f9c7f24f3)
- locally with no authentication, using the CLI as trigger