Avoid active settings bleed between unit tests (#719)
This fixes the bleeding of mock settings between unit tests.

It was discovered during debugging of #714 that mock settings that were
set up by unit tests wound up in the settings of other subsequent unit
test executions, if the latter did not explicitly override them.

This due to the fact that the fixture function used to mock the settings
replaced the global raw settings cache, but did not replace the global
active settings cache, which also contained env var overrides.

This PR updates the mock settings fixture function to replace both
global settings caches, and also refactors it in order to improve
readability.