Fix S3Location Dashboard link generation for directories (#725)
The links generated in the Dashboard for `S3Location`s that pointed to
"directories" rendered an error page.

This is due to S3 only emulating a file system through key-value pairs,
without having an actual hierarchical directory structure. S3 has 2
separate endpoints for "directories" and files: "buckets" and "object",
respectively

The existing Dashboard hyperlink generation mechanism generates these
links using only the "object" endpoint, which results in an error being
rendered for "directories".

This PR updates the generation of hyperlinks to S3 in the Dashboard to
use the "buckets" endpoint for all locations. In this case, S3
successfully redirects to the "object" endpoint for existent files.
Rendering of file links is unaffected.

This PR also introduces a convention in which locations that end in a
"/" are considered "directories", and those that do not are considered
as fully-qualified file paths.

A Testing Pipeline capability of composing S3 links is added, and unit
tests are updated to enforce the new convention.

```bash
$ bazel run sematic/examples/testing_pipeline:__main___local -- \
    --s3-uris s3://<redacted dir path>/ s3://<redacted dir path>/<redacted file name>
```


![image](https://user-images.githubusercontent.com/1894533/230690908-e4fe82fe-3d12-4ed6-817d-fe0c7d84cd8e.png)

**Before:**

![image](https://user-images.githubusercontent.com/1894533/230690118-58b54090-14ac-4932-8ceb-ea8f4369c5cc.png)

**After:**

![image](https://user-images.githubusercontent.com/1894533/230690500-d46a6a9f-cf5a-420e-b7e9-5540be4246e7.png)