Implements new style Notes components (#708)
Adds new components of the right column of the Rundetails page, which
includes a note submission component and notes list.

A new Storybook page is created for the two components mentioned above. 

Functionalities:

1. NoteSubmission component automatically increases its space if a user
types multiple lines of text.
2. Note component supports folding up its contents if too long and can
be expanded. The folding control button is hidden if the content is
short enough.
3. The name of the author of the note will be shortened with an ellipsis
if it's too long. A tooltip is added only if the name is too long.

The result can be previewed here: 


https://story.dev-usw2-sematic0.sematic.cloud/iframe.html?args=&id=sematic-page--page&viewMode=story


![capture1](https://user-images.githubusercontent.com/1046489/229419027-1984585f-39f4-4210-b126-792e7f6427ab.gif)