Make gevent monkey patch occur early (#728)
When using gevent, it requires a monkey patch to some standard
libraries. This monkey patch has to be applied *very early*, before
certain other stdlib imports occur (even in transitive deps). The import
is basically only required if you are using the server, but if the patch
is applied at the wrong time when doing other things it can cause
strange behaviors in other libs. This PR makes it so that we install a
[`.pth` file](https://docs.python.org/3/library/site.html) with our
wheel which will apply the patch before the user's main module is
loaded. The `.pth` file will only work when installed with the wheel
though, so some changes are also added in this PR to do an early patch
for running the server and the CLI as main for when we're working
directly in bazel.

Testing
-------
- built wheel, started local server and ran a pipeline against it using
python 3.10
- used bazel with `bazel run //sematic/cli:main_py3_10 -- start` to
start the server and ran a pipeline against it
- used `serve-dev` to deploy and ran a pipeline