Enable reverse log traversal (#718)
Before this PR, you could only start from the beginning of logs and go
all the way to the end by loading everything in between. This PR makes
it so you can *also* start at the end, and go either forward or backward
from there. From any log line result, you have a cursor for the forward
direction to traverse more in that direction, or a reverse cursor to
traverse more in that direction.

Testing
-------
Executed the testing pipeline with log spamming and sleeping that would
both traverse multiple pages of logging. Also modified the inline
function to do multiple pages of logging. Traversed logs starting from
the beginning and going to the end as well as starting at the end and
going to the beginning for each of these (hacked the flask endpoint to
enable the reverse direction for testing purposes; eventually the
front-end will need controls for this). Checked around page breaks to
ensure there were no missing/duplicated lines due to off-by-one errors.