Add line ids for log lines (#733)
The front-end wants a more robust way to ensure that duplicate lines are
not shown in the log UI (due to duplicated requests, for example). This
PR adds unique line ids for each returned log line that is derived from
the log line's source file and index in such a way as to ensure that the
ids are unique within a given run, and that higher ids are for later log
lines.

Testing
-------
In addition to unit tests, also deployed and confirmed that the line ids
are being returned by the HTTP request for log lines.