Refactors devLogger creation (#739)
The `useLogger` is used in many React components; each time a component
consumes it, it has to re-fetch the setting in the URL and local
storage, which is not performant.

1. The debug flag is set per page load and persists unless the page is
reloaded again. The setting does not need to be re-read when the page
session is live.

2. The logging function instance can be memorized instead of re-created
for each component.

This PR addresses the above issues by leveraging [lodash's
memoize](https://www.geeksforgeeks.org/lodash-_-memoize-method/)