Force HTTP requests to redirect to HTTPS by default (#748)
accidental usage of `http://` urls that are insecure, and cause google
oauth to throw errors because it does not allow mixed use of
secure/insecure domains. rather than failing silently, we set all http
requests to redirect to https by default so that users have a secure
default. also added `ingress.force_ssl` as an option so that this can be
disabled for local/unsecured clusters.

fixes #597