Add Sematic Grafana dashboards as helm package (#767)
creates a separate helm package called `sematic-grafana-dashboards`,
that will be served up by the same helm release mechanism we currently
have in the `sematic-ai/helm-charts` repo, i.e. it will share the same
`index.yaml` file in that repo, but will be it's own standalone helm
package that must be installed separately by the user.