Migrate shared code to common (#776)
Migrate `useLogger` and `useHttpClient` to the `common` package, so both
the `common` and `main` packages can use them. (We forbid `common`
package referencing stuff from `main`, the opposite is fine)

Tips for review: most of changes to the files in the `main` package are
import path changes. As long as the `main` package` compiles
successfully, there is not much need to review them.