Fix cleaning of runs in the CREATED state (#791)
Runs can't transition from CREATED to FAILED/NESTED_FAILED. The
assumption is that the run would not have had a chance to fail if it was
never even scheduled (and therefore also had no children). Thus to move
it directly to a terminal state, we must move them to CANCELED. This PR
updates the cleaning of CREATED orphaned runs to move them to CANCELED.

Testing
-------
In addition to unit tests, deployed in my namespace (where I was seeing
some errors related to this and some old orphaned runs) and confirmed
that some runs which had not previously been cleanable got cleaned up.