[3/N] Docker Build System unit tests (#807)
This is the third PR in a chain that introduces a Build System plugin
with an implementation that uses native Python+Docker capabilities to
build a container image and launch a pipeline, without requiring Bazel.
The previous PR in the chain is #790.

This PR introduces unit tests that cover the functionality from the
first two PRs:

```bash
$ bazel test --test_output=all sematic/plugins/building/tests:test_docker_builder
```