[4/N] Build System configurable Docker connection (#813)
This is the third PR in a chain that introduces a Build System plugin
with an implementation that uses native Python+Docker capabilities to
build a container image and launch a pipeline, without requiring Bazel.
The previous PR in the chain is #807.

This PR makes the Docker Client connection configurable, and fills in
docstrings.

## Testing

Validated that the intermediate and advanced cases from the
[`example_docker` repo](https://github.com/sematic-ai/example_docker)
still work, and added unit tests and checks for the new funcitonality:

```bash
$ bazel test --test_output=all sematic/plugins/building/tests:test_docker_builder
```