Add horizontal pod autoscaling and pod disruption budget for API server (#811)
fixes #743.

all changes listed below apply only to the api server. socket io server
is unaffected, disruptions on it are considered "acceptable" since it is
not mission critical for correctness.

- increases default replica count to 2. this is required for fault
tolerance and so that the pod disruption budget of 1 does not prevent
nodes from draining.
- adds horizontal pod autoscaling (hpa) support, can be optionally
enabled
- adds pod disruption budget of min 1 pod available, enabled by default
- hpa is now supported via the latest kube api version `autoscaling/v2`
- server cpu/memory resource request/limits are both halved to prior
values, to account for 2 replicas being deployed by default compared to
1 before.

helm doc changes to follow in a separate pr.