[6/N] Extract and reuse utility functions in DockerBuiler (#817)
This is the sixth PR in a chain that introduces a Build System plugin
with an implementation that uses native Python+Docker capabilities to
build a container image and launch a pipeline, without requiring Bazel.
The previous PR in the chain is #815.

This PR eliminates tech debt by extracting two local functions from
other code modules to utility functions, and reusing them in the new
Docker Build System, in addition to the original usages.