Code clean up (#819)
A few things in this PR:
1. Migrate `MuiRouterLink` to `common` package, so `common` package can
use it. It is important to use this component to avoid full page reload
when clicking links. References are updated.
2. Migrate `ExtractContextType` to `common` package. This is merely a
typing helper. References are updated.
3. Fixes an issue: missing `endedAt || failedAt!` fields from a canceled
run might cause a page crash.
4. Using
[setupProxy.js](https://github.com/sematic-ai/sematic/compare/c/code_cleanup?expand=1#diff-9d1df0e81c02e81de08e2568a76d323e716b5537b4b7c6d1f2a07d01b4de1e70)
instead of `proxy` field in package.json. The benefit of `setupProxy.js`
is that it can read env vars in the code logic. So we can export local
override in Shell without always changing `package.json` and revert it
before committing (A nightmare for those who frequently work in the FE).
This can also be a stepping stone for future mocked server work. More
details can be found
[here](https://create-react-app.dev/docs/proxying-api-requests-in-development/#configuring-the-proxy-manually).