Have resolver continue in a new resolution if resolver restarts (#801)
It is possible for the resolver to restart mid-execution. When this
happens, we ideally would be able to have the resolver pick back up
where it left off (see #141). However, in the meantime while we
determine the best way to make that happen, we can just have the
resolver start a new resolution continuing from where the current one
left off. This PR implements that behavior.

Testing
--------
In addition to unit test, I also launched this
[resolution](https://josh.dev-usw2-sematic0.sematic.cloud/runs/726af2c8bc484b468270dd5007dfd2c0#tab=output&panel=run)
and manually forced the resolver pod to get rescheduled mid-resolution.
It created this [new
resolution](https://josh.dev-usw2-sematic0.sematic.cloud/runs/07aa13313fab493891f4ac9ae38fd331#tab=output&panel=run)
as expected and referenced it in the error message for the original.

---------

Co-authored-by: Josh Bauer <josh@sematic.dev>