Adjust the Name column width distribution (#829)
The PR covers:

1. On the run search page, if the Name column width is shrunk,
prioritize shrinking the width of the import path instead of squeezing
the run name. So there is a better chance to see the complete run name.
However, if the column width is further reduced, part of the run name
remains hidden. Illustrated below:


![capture1](https://github.com/sematic-ai/sematic/assets/133257643/414f7135-4ec7-40f8-adf3-6101b781ce73)

2. Introduced `<ErrorBoundary />` component, so errors will not bubble
up and destroy the whole page.

3. Improved the rendering logic of run status logic; there are now more
default/fallback values for extreme cases.

4. Fixed an issue in the local dev login process, which didn't proxy the
`login` request.

The change can be previewed in `my` namespace.

Co-authored-by: Chance An <anchangsi@gmail.com>