Add support for searching by tags (#831)
This PR adds support for searching for runs via tag filters. It follows
the same structure for search filters as the existing filters, but adds
a new operator `"contains"` in addition to `"eq"` and `"in"`. Since
different column types may implement `"contains"` differently, a
mechanism is added to allow columns to specify how to perform
`"contains"` searches with them.

Testing
-------
In addition to unit tests, deployed to my namespace and tried some
queries manually using `requests`.