Fix Docker image tag name reload (#833)
When using the Native Docker Build System to build and push container
images to a remote repository, if the image had already been built,
tagged, and pushed, reloading its remote name and digest from the
repository results in using the local name instead of the remote name.

This is due to using a wrong index to refer to the image within a
`docker-py`-specific attribute property. There does not seem to be an
API to get a hold of the new image name through a more
officially-supported way.

This PR fixes the issue by always referring the last-added image name
among the image attributes.