Enable run search deep link (#845)
The GitHub integration feature requires that we be able to link to a run
list filtered to only include runs tagged with `checks-commit:<some
commit id>`. We don't currently have the ability to link to any
particular search filter. This PR adds support for that, in ~both the
new and~ the existing run list page. Only the "search" filter is
supported, as that filter is present in both variants of the UI. That
means the GitHub integration can link to a raw text search for
`checks-commit:<some commit id>` and have the link work the same
regardless of whether the user has the new or old UI enabled.

**Edit**: At Chance's request, have removed the changes to the new UI.

Testing
-------
For ~both the new and~ the old UI:
- confirmed that visiting the run list page with no URL query parameters
works as it already does
- confirmed that adding the `?search=<some search text>` query param
prepopulates the search box and displays run search results appropriate
to that filter
- confirmed that you can still enter and use user-specified filters even
if the URL still contains `?search=<some search text>`