[7/N] Relative build config paths and config file merging (#850)
This is the seventh PR in a chain that introduces a Build System plugin
with an implementation that uses native Python+Docker capabilities to
build a container image and launch a pipeline, without requiring Bazel.
The previous PR in the chain is #817.

This PR switches from project root-relative paths to build config
file-relative paths. This means that all the configurations values which
represent file paths or globs inside build configuration files will now
be considered relative to the respective configuration files.

The paths are also validated to represent real files and to not attempt
to reference files outside of the project file structure.

It is possible to still specify files relative to the project root by
prefixing the path with `"//"`.

This PR also implements build config file merging down the project file
structure chain. When a launch script is executed, files with the name
`"sematic_build.yaml"` are searched for starting from the current
working directory and down to the directory in which the launch script
is located. In addition, a build config file with the name of the launch
script, but with the `.yaml` extension is also loaded, if present. These
configurations are then loaded and merged, with configurations closer to
the launch script overriding or extending configuration values specified
earlier.

## Testing

Unit and integration tests are added to validate configuration loading
and merging, and file path normalization and validation:

```bash
$ bazel test --cache_test_results=no sematic/plugins/building/tests:test_docker_builder_config
```

The `example_docker` project is updated in
[#2](https://github.com/sematic-ai/example_docker/pull/2) to also use
build config file-relative paths.