Add build config files for Ray example pipelines (#857)
This is a follow-up to #844 that adds build configuration files to the
example pipelines which require `Ray` as a dependency.

This is enabled by adding the source code of the example pipelines which
depend on ray to the "extra" wheels.

## Testing

Because these pipelines now depend on the version of `sematic[ray]` that
is built using these changes, which are not released to PyPi, they were
tested using a local build of the Sematic wheel. [This
commit](https://github.com/sematic-ai/sematic/pull/857/commits/46ae583f0d0656f77e0dfbe8f027ba1e43776418)
contains the version that was tested.

**Example Testing Pipeline** run command:

```
$ sematic run --build sematic/examples/testing_pipeline/__main__.py -- --images --cloud --detach
```

Resulting Dashboard entry:
[link](https://dev1.dev-usw2-sematic0.sematic.cloud/runs/8813fdd626c042c8bf03e1500cf884e8#tab=output&panel=run&run=3241e3f3628d455b8fe8459380d1c423).

**Example Lightning ResNet** run command:

```
$ sematic run --build sematic/examples/lightning_resnet/main.py -- --cloud
```

Resulting Dashboard entry:
[link](https://dev1.dev-usw2-sematic0.sematic.cloud/runs/1f75060f92bf40dd91f12eb042af221e#panel=run&tab=logs&run=0ed5f1179b144a629e8e7faf09c78051)
(did not complete due to [this
error](https://github.com/sematic-ai/sematic/pull/857#issuecomment-1572144611))

**Example CIFAR10 Classifier** run command:

```
$ sematic run --build sematic/examples/cifar_classifier/main.py -- --cloud
```

Resulting Dashboard entry:
[link](https://dev1.dev-usw2-sematic0.sematic.cloud/runs/3f02586fd29a4839adfbbe36d9f5edf5#panel=run&tab=logs&run=34c16ddf0e0746188554caa6a047ba68)
(did not complete due to [this
error](https://github.com/sematic-ai/sematic/pull/857#issuecomment-1572144611))